LayoutTests/imported/w3c/web-platform-tests/xhr/overridemimetype-edge-cases.window-expected.txt


FAIL overrideMimeType() state needs to be reset across requests assert_equals: expected "Âð" but got "\ufffd\ufffd"
PASS If charset is not overridden by overrideMimeType() the original continues to be used 
FAIL Charset can be overridden by overrideMimeType() with a bogus charset assert_equals: expected "\ufffd\ufffd" but got "Âð"
